+++ tests/data/snapshots/linux/firefox/myAccountPage.spec.ts/default.png
