+++ tests/data/snapshots/win32/chromium/myAccountPage.spec.ts/default.png
