+++ tests/data/snapshots/win32/firefox/myAccountPage.spec.ts/default.png
