+++ tests/data/snapshots/win32/webkit/myAccountPage.spec.ts/default.png
